Renamed base class and package
