+++ pom.xml
@@ -12,9 +12,9 @@
     </parent>
 
     <groupId>com.exadel</groupId>
-    <artifactId>sandbox-team3-be</artifactId>
+    <artifactId>discount-web-app</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>sandbox-team3-be</name>
+    <name>discount-web-app</name>
     <description>Demo project for Spring Boot</description>
 
     <properties>
