+++ src/main/java/com/exadel/discountwebapp/controllers/HelloController.java
@@ -1,4 +1,4 @@
-package com.exadel.sandboxteam3be.controllers;
+package com.exadel.discountwebapp.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
